Advert provides getters for its fields
Model should probably be rethought, but for that Scala must be better investigated. Seems that abstract class + case classes extending it is not a great approach. Maybe case classes that inherit from others should be restricted for pure algebraic types? For the time being, getters are added until further investigation.